Separate modules for StyleSheet and related side-effects
